# Improve link parser performance (mostly by removing `enum`)